+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.0
 
 - Add a post install message
+- `bundle download clear` command to remove downloads (alias as `clean` too)
+- Support `os` option for `download` DSL keyword in `Downloadfile`
 
 ## 1.0.0